Give more clearer error message for missing scope.
Before, most views would treat a request
authenticated with client credentials as a request
authenticated with non client credentials if the
full scope was not present (even though other
scopes were present).

This would lead to an error because it would
check which groups the user is part of, but when
a request is authenticated with client
credentials, it does not have any groups (and in
fact, the user is not created in the database
because it is only necessary for the request).

Now, the check is done if any scopes are present,
and if only if there are no scopes, it's assumed
that it's not authenticated via client
credentials.

In the other case (there are scopes but no full
access scope), that's actually logged as the
error, so debugging is easier (e.g. the
SCOPE_FULL_ACCESS setting has to be configured.